+++ force_wfmanager/central_pane/configuration.py
