+++ force_wfmanager/left_side_pane/plugin_manager.py
